// src/app/communication-protocol/model/protocol-param-value.ts
export class ProtocolParamValue {
    id: number;
    commProtocolParamId: number;
    value: string;
    refCommProtocolParamValueId: number;
    isDefaultValue: boolean;
    status: string;
    createdBy: number;
    updatedBy: number;
    isDefalutValueString: string;
    refCommProtocolParamValueName: string;
  }